google-ads-automation/requirements.txt
 google-ads
 google-auth
 beautifulsoup4
 requests
 sys
 os
 